Rename py-image-builder to pymage (module, imports, CLI, workflow, docs)
