+++ pymage/internal/lock/lock_test.go
