+++ pymage/internal/ptar/ptar_test.go
