+++ pymage/internal/wheel/wheel.go
@@ -21,7 +21,7 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/imjasonh/terraform-playground/py-image-builder/internal/ptar"
+	"github.com/imjasonh/terraform-playground/pymage/internal/ptar"
 )
 
 // Layout describes where files are installed inside the image filesystem.
@@ -181,7 +181,7 @@ func (w *Wheel) consoleScripts(bin string) ([]ptar.File, error) {
 	for _, n := range names {
 		module, attr := eps[n].module, eps[n].attr
 		script := fmt.Sprintf(`#!/usr/bin/env python
-# Generated by py-image-builder.
+# Generated by pymage.
 import sys
 from %s import %s
 if __name__ == "__main__":
